+++ ウィンドゥ表示.vcxproj.filters
@@ -24,7 +24,7 @@
     <ClCompile Include="main.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
-    <ClCompile Include="2D\Mesh.cpp">
+    <ClCompile Include="2D\Sprite.cpp">
       <Filter>ソース ファイル\2D</Filter>
     </ClCompile>
   </ItemGroup>
@@ -36,7 +36,7 @@
     <None Include="Basic.hlsli" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="2D\Mesh.h">
+    <ClInclude Include="2D\Sprite.h">
       <Filter>ヘッダー ファイル\2D</Filter>
     </ClInclude>
   </ItemGroup>
